Woodog/requirements.txt
six==1.15.0
Django==3.0.7
django_heroku==0.3.1
validate_email==1.3
asgiref==3.3.4
dj-database-url==0.5.0
Django==3.0.7
django-filter==2.4.0
django-heroku==0.3.1
djangorestframework==3.12.4
Markdown==3.3.4
psycopg2==2.9.1
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
validate-email==1.3
whitenoise==5.2.0
